divide tests into integration and unit
* also migrate test utils into a utils mod
